hackathon_day/requirements.txt
paho-mqtt
pygame
numpy
vqf
